backend/requirements/base.txt
#DJANGO
Django~=4.1.5
psycopg2~=2.9.5
django-environ~=0.9.0
djangorestframework~=3.14.0
drf-spectacular~=0.25.1
django-redis~=5.2.0
celery[redis]~=5.2.7
uvicorn[standard]~=0.20.0
gunicorn~=20.1.0
 
